+++ lib/src/functions.dart
@@ -793,6 +793,92 @@ class Abs extends DefaultFunction {
 }
 
 /**
+ * The ceil function.
+ */
+class Ceil extends DefaultFunction {
+
+  /**
+   * Creates a new ceil function with given argument expression.
+   */
+  Ceil(arg): super._unary('ceil', arg);
+
+  /// The argument of this ceil function.
+  Expression get arg => getParam(0);
+
+  /// Ceil never has a slope.
+  //TODO No differentiation possible for integer x
+  Expression derive(String toVar) => new Number(0);
+
+  /**
+   * Possible simplifications:
+   *
+   * 1. ceil(floor(a)) = floor(a)
+   * 2. ceil(ceil(a)) = ceil(a)
+   */
+  Expression simplify() {
+    final Expression sarg = arg.simplify();
+    return sarg is Floor || sarg is Ceil? sarg : new Ceil(sarg);
+  }
+
+  evaluate(EvaluationType type, ContextModel context) {
+    var argEval = arg.evaluate(type, context);
+
+    if (type == EvaluationType.REAL) {
+      return argEval.ceil();
+    }
+
+    if (type == EvaluationType.VECTOR) {
+      //TODO apply function to all vector elements
+    }
+
+    throw new UnimplementedError('Can not evaluate ceil on ${type} yet.');
+  }
+}
+
+/**
+ * The floor function.
+ */
+class Floor extends DefaultFunction {
+
+  /**
+   * Creates a new floor function with given argument expression.
+   */
+  Floor(arg): super._unary('floor', arg);
+
+  /// The argument of this floor function.
+  Expression get arg => getParam(0);
+
+  /// Floor never has a slope.
+  //TODO No differentiation possible for integer x
+  Expression derive(String toVar) => new Number(0);
+
+  /**
+   * Possible simplifications:
+   *
+   * 1. floor(floor(a)) = floor(a)
+   * 2. floor(ceil(a)) = ceil(a)
+   */
+  Expression simplify() {
+    final Expression sarg = arg.simplify();
+    return sarg is Floor || sarg is Ceil? sarg : new Floor(sarg);
+  }
+
+  evaluate(EvaluationType type, ContextModel context) {
+    var argEval = arg.evaluate(type, context);
+
+    if (type == EvaluationType.REAL) {
+      return argEval.floor();
+    }
+
+    if (type == EvaluationType.VECTOR) {
+      //TODO apply function to all vector elements
+    }
+
+    throw new UnimplementedError('Can not evaluate floor on ${type} yet.');
+  }
+}
+
+/**
  * The sign function.
  */
 class Sgn extends DefaultFunction {